Refactors entity components into a manager
Moves entity stats and components into a ComponentManager node.
This change provides a more structured and modular approach to managing
entity components, improving code organization and maintainability.
It also allows easier access to components and stats.
The old StatView has been removed as it is no longer needed, it is
replaced with a new StatView that takes a stat_id to be more generic.