Fix host player unable to cast spells in 'local_main'
When the EntityManager is used to 'spawn()' something, it will call some Global signals for setup purposes. Some of the child nodes in the Entity scene are dependent on those signals.

If the Entity is not readied in the SceneTree yet, then calling 'EntityManager.spawn()' will result in those child nodes not getting wired up. Further, because the Entity is registered in the EntityManager as being spawned already (via being assigned an ID), when the Entity does enter the SceneTree and calls the 'spawn()' method itself it will short-circuit resulting in the child nodes being unable to ever get wired up.

The easiest example of this (and the way I noticed) was casting the '2' spell to throw the projectile. This simply didn't work on the Host player, but worked fine for joining players. Best I can tell, this is because the MultiplayerSpawner node was syncing the SceneTree state of the level root. So when a client would "Join", they would sync that state which included an Entity in it already. This in turn would result in things working correctly since the Entity was already live in the tree before calling the 'EntityManager.spawn()'. Keep in mind this isn't true server/shard data replication (in a weirdly technical sense, these Entities are actually the same player character), so this is largely just the nature of this particular test scene.

PS. This is just a testing scene, but worth noting this behavior when we get around to true replication. May need a more robust way of ensuring child dependencies in scenes can get wired up correctly on their own without relying on this order of operations.