Add interactables (#176)
# What

Taking a first pass at interactables. High-level changes are:

- Interact Keybind: 'F' (M&K) or 'X/Square' (Controller)
- Added a general "interaction" system
  - InteractionManager for global management
  - InteractableComponent for making an Entity interactable
  - InteractionComponent for enabling an Entity (Player) to interact with interactables
  - Various helper classes for supporting the above
- Updated the devmap scene with some test interactables
  - Door: Opens immediately on interact
  - Pick up: Is "picked up" after 2 seconds
  - Control Zone: Player can stand in the zone to interact with it, completes after 1 second

# Why

Well we gotta let the player interact with stuff! Interactive experiences are way more fun when you can interact with and experience things.

# Extra
This is still very much in a zombie state. It needs a lot of cleanup, but is technically functional. Most importantly, this system will likely need a lot of changes once we added a Targeting system. Implementing the Interaction system while trying to keep it open for the Targeting system, UI, etc. was a bit tricky, but I think I've got something functional at least.

Fully expect this to get updated over time. It's overly simple in some places and overly complex in others.