Refactors entity component system
Migrates to a more flexible component system.

Removes the `ComponentManager` class and replaces it with a more generic node-based system. This allows for easier addition and removal of components at runtime, and simplifies the code.

Renames and moves various scripts to better reflect their purpose.

Removes the `EntitySpawner` scene as it is no longer needed.

Adds a new `Projectile` entity scene.