Add interactable requirements (#191)
# What

Added optional "requirements" to Interactable objects. When applied, these requirements determine IF a given Entity can interact with that Interactable. If the requirement's condition passes, the interaction is "allowed" and will go through when attempted.

# Why

Allows us to restrict interaction with objects by arbitrary conditions. Especially now that the player can target objects and subsequently interact with their current target. For example, if the player targets a "door" and tries to interact from a million miles away, the interaction probably shouldn't be allowed. Instead the door should enforce how "close" the player should be for the interaction to succeed.


PR: https://github.com/Uraxii/naous/pull/191