Base enemy logic started with example implementation
# What

Add a base enemy class and one empty implementation example. The enemy is built via the same Entity pattern as the Player is. In its current form, the enemy includes:

- Entity sync nodes
- Body and mesh nodes (composable and replaceable)
- Components for health, speed, gravity, targetable
- NEW: Player Detection component. Draws a ray that always points towards the player with a provided length. If the player intersects that ray (ie. gets close enough to the enemy), the player will be "detected".
- NEW: Component for AI code-driven movement. Currently just moves at a constant speed towards the player when the player is detected and doesn't move at all when not detecting a player.

# Why

To set up a baseline for enemy entities that we can build on for later. This is a very minimalist implementation and is built with the expectation that it will need changed later.