Completed Tutorial Sequencing (#263)
# What

Added the Backup and Escape sequences to the tutorial. This completes the core structure of the tutorial, though it still needs a good amount of feature adds and polish to feel player-ready.

# Why

In completing this sequencing, we should now be able to jump into the `TutorialSequence` class and make tweaks to the various steps within the sequence as necessary. The critical path that moves the sequence forward is done (and we can add more steps as necessary) so now we just need to fill it in with some goodies!